Moving BatchIdToEntityHelper to DTO namespace which is excuded from autowiring
